==> r/03_traitement.Rmd <==
---
title: "03_traitement"
author: "Psqrt"
date: "16/12/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Packages et fonctions externes

```{r, message=F, warning=F}
source("./0_packages.R")
```

```{r}
type_col = readRDS("./../data/data_sel_type_col.rds")
data_sel = data.table::fread(file = "./../data/data_sel.csv", colClasses = type_col) %>% as.data.frame()
```

## AFDM

```{r}
df<- data_sel %>% filter(dataset=="train")
if (0) {
    res_famd <- FAMD(df[-c(1:3)], graph = FALSE)
    saveRDS(object = res_famd,
            file = "./../data/res_famd.rds")
}
```


```{r}
res_famd<-readRDS("./../data/res_famd.rds")
coord<-res_famd$quali.var$coord %>% as.data.frame()
liste_var_factor = type_col[type_col == "factor"]
liste_var_factor<-names(liste_var_factor[-1])
name<-c()

for (var in liste_var_factor) {
    df<-data_sel %>% filter(dataset=="train") %>% select(var) 
    colnames(df)<-"col"
    n<-length(levels(df$col))
    name<-c(name,rep(var,n))
}
name <-name[-c(239,302)]
# On retire une de chaque a cause de modalite non presente dans train (ps_calc_06 et ps_calc_11)
coord<-cbind.data.frame(name,coord) 

modalite<-data.frame("modalite"=rownames(coord), stringsAsFactors = FALSE) %>%
    mutate(nb=str_count(modalite, pattern = "\\."), clean = case_when(nb == 0 ~ gsub("X","",modalite),
                                                                      nb == 1 ~ str_extract(modalite, regex("(?<=X)[0-9]*")),
                                                                      nb == 2 ~ str_extract(modalite, regex("(?<=X)[0-9]*\\.[0-9]*"))))
coord<-cbind.data.frame(coord,"modalite"=modalite$clean)
coord$modalite<-as.character(coord$modalite)
coord$modalite[which(name=="ps_reg_01")]<-seq(0,0.9,0.1)
coord$modalite[which(name=="ps_reg_02")]<-seq(0,1.8,0.1)
for(var in liste_var_factor) {
    df<- coord %>% filter(name==var) %>% select(-name)
    print(gg_proj_ind(df,1,2) + geom_text_repel(aes(label=modalite)) +ggtitle(var))
}
```


## Tables croisées avec la target

```{r}
liste_var_factor = type_col[type_col == "factor"]
```

```{r}
for (var in names(liste_var_factor)[2:length(liste_var_factor)]){
    print(gg_contingence_freq(get("var")))
}
```

Remarque : régression logistique sur ps_car_11_cat, on récupère les beta qui serviront de "modalités" pour construire une variable continue (faire un join ensuite pour remplacer les modalités par les beta).

* `ps_ind_01` : modalité 6 vers 4, modalité 7 vers 3 ou 5
* `ps_ind_03` : modalité 0 vers ? (peut être la garder), modalité 9 vers 1, modalité 10 vers 8, modalité 11 vers 8
* `ps_ind_15` : modalité 1 vers 0, modalité 2 vers 0
* `ps_reg_01` : modalité 0.0 vers 0.7 ou 0.8, modalité 0.2 vers 0.5
* `ps_reg_02` : modalité 0.0 à 1.0 ensemble, 1.1 à 1.8 ensemble
* `ps_car_06_cat` : modalités 2, 5, 8, 9, 12, 13, 15, 17 ensemble, modalités 0, 1, 3, 4, 6, 7, 10, 11, 14, 16 ensemble
* ` ps_car_11_cat` : nah
* `ps_calc_04` : modalité 0 vers 1 ou 2 ou 3 ou 4, modalité 5 atypique
* `ps_calc_05` : modalité 5 vers tout possible, modalité 6 très atypique
* `ps_calc_06` : modalités 0 à 5 ensemble, modalités 6 à 10 ensemble
* `ps_calc_07` : modalités 0 à 3 ensemble, modalités 4 à 9 ensemble
* `ps_calc_08` : modalités 2 à 7 ensemble, modalité 12 vers 11
* `ps_calc_09` : modalités 4 à 7 ensemble
* `ps_calc_10` : modalités 0 à 4 ensemble, modalités 13 à 25 ensemble
* `ps_calc_11` : modalités 0 à 2 ensemble, modalités 9 à 19 ensemble
* `ps_calc_12` : modalités 4 à 10 ensemble
* `ps_calc_13` : modalités 6 et 13 ensemble
* `ps_calc_14` : modalités 0 à 3 ensemble, modalités 12 à 23 ensemble
* `ps_car_01_cat` : modalités 0 à 5 ensemble, modalités 8 et 9 ensemble


```{r}
levels(data_sel$ps_ind_01) = c("A", "B", "C", "D", "E", "F", "G", "G")
levels(data_sel$ps_ind_03) = c("A", "A", "B", "C", "D", "E", "F", "G", "H", "I", "I", "I")
levels(data_sel$ps_ind_15) = c("A", "B", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M")
levels(data_sel$ps_reg_01) = c("F", "A", "A", "B", "C", "D", "E", "F", "G", "H")
levels(data_sel$ps_reg_02) = c("A", "B", "C", "D", "E", "F", "G", rep("H", 12))
levels(data_sel$ps_car_06_cat) = c("A", "B", "C", "D", "E", "F", "D", "D", "C", "F", "G", "H", "D", "C", "I", "C", "F", "F")
levels(data_sel$ps_calc_04) = c("A", "A", "B", "C", "D", "D")
levels(data_sel$ps_calc_05) = c("A", "B", "C", "D", "E", "E", "E")
levels(data_sel$ps_calc_06) = c("A", "A", "A", "A", "A", "A", "A", "B", "C", "D", "E")
levels(data_sel$ps_calc_07) = c("A", "A", "B", "C", "D", "E", "E", "E", "E", "E")
levels(data_sel$ps_calc_08) = c("A", "A", "A", "A", "A", "A", "B", "C", "D", "E", "E")
levels(data_sel$ps_calc_09) = c("A", "B", "C", "D", "E", "E", "E", "E")
levels(data_sel$ps_calc_10) = c("A", "A", "A", "A", "A", "B", "C", "D", "E", "F", "G", "H", "I", rep("J", 13))
levels(data_sel$ps_calc_11) = c("A", "A", "A", "B", "C", "D", "E", "F", "G", rep("H", 11))
levels(data_sel$ps_calc_12) = c("A", "B", "C", "D", rep("E", 7))
levels(data_sel$ps_calc_13) = c("A", "B", "C", "D", "E", "F", rep("G", 8))
levels(data_sel$ps_calc_14) = c("A", "A", "A", "A", "B", "C", "D", "E", "F", "G", "H", "I", rep("J", 12))
levels(data_sel$ps_car_01_cat) = c("A", "A", "A", "A", "A", "A", "B", "C", "D", "D", "E", "F")
```


## Régression logistique sur variable à 104 modalité

 
```{r}
df <- data_sel %>% filter(dataset=="train")
model<-glm(data = df, target ~ ps_car_11_cat, family = binomial(link = logit))
res<-summary(model)
df<-cbind.data.frame(ps_car_11_cat=gsub("ps_car_11_cat","",rownames(res$coefficients)),res$coefficients)
df<- df %>% select(ps_car_11_cat,Estimate)
df$ps_car_11_cat<-as.factor(df$ps_car_11_cat)
data_sel<- left_join(data_sel,df,"ps_car_11_cat")
data_sel$ps_car_11_cat<-data_sel$Estimate
data_sel <- data_sel %>% select(-Estimate)
ggplot(data_sel, aes(ps_car_11_cat,color=target))  +
  geom_density() +ggtitle("ps_car_11_cat")
```


## Transformation des variables continues



## Exportation

Après avoir transformé les deux échantillons, on peut les rediviser selon la colonne "dataset". On exportera deux bases dans ce cas (ne pas oublier de récupérer le type des colonnes et exporter ça également).


```{r}
train<-data_sel %>% filter(dataset=="train") %>% select(-dataset)
test<-data_sel %>% filter(dataset=="test") %>% select(-dataset)

type_col = sapply(train, class)

saveRDS(object = type_col,
      file = "./../data/type_col.rds")


write.csv(x = train,
          file = "./../data/train.csv",
          quote = T,
          row.names = F)
write.csv(x = test,
          file = "./../data/test.csv",
          quote = T,
          row.names = F)
```
