==> r/4_correlations_et_dependances.Rmd <==
---
title: "Études des corrélations et dépendances"
date: "Étape 4"
output:
  rmdformats::readthedown:
    highlight: kate
---

```{r knitr_init, echo=FALSE, cache=FALSE}
library(knitr)
library(rmdformats)

## Global options
options(max.print = "75")
opts_chunk$set(echo    = TRUE,
               prompt  = FALSE,
               tidy    = FALSE,
               comment = NA)
opts_knit$set(width = 75)
```

# Importation des Packages et fonctions externes

```{r, message=F, warning=F}
source("./0_packages.R")
```

# Importation des données

```{r}
type_col = readRDS("./../data/data_sel_type_col.rds")
type_col = type_col[-1]
train = fread(file = "./../data/train.csv", colClasses = type_col) %>% as.data.frame()
test = fread(file = "./../data/test.csv", colClasses = type_col) %>% as.data.frame()
```

## Étude des corrélations entre variables explicatives continues

```{r}
var_num = names(train)[sapply(train, class) == "numeric"][-1]
var_fac = names(train)[sapply(train, class) == "factor"][-1]

train %>% 
    select(var_num) %>% 
    cor() %>% 
    corrplot(method = "number",
             type = "upper",
             tl.cex = 0.8,
             tl.srt = 45,
             tl.col = "black")
```

## Étude ANOVA entre variables explicatives continues et la target

```{r}
df_recap_anova = data.frame("variable" = var_num, "pvalue" = NA, stringsAsFactors = F)
for (i in var_num){
    print(i)
    aov_result = summary(aov(get(i) ~ target,
                data = train))
    print(aov_result)
    df_recap_anova[which(df_recap_anova$variable == i), "pvalue"] = aov_result[[1]][["Pr(>F)"]][1]
}
df_recap_anova %>% arrange(-pvalue)
```

On note que 6 variables ne passent pas l'ANOVA. Ce sont essentiellement des variables type `calc` : `ps_calc_11`, `ps_calc_14`, `ps_calc_02`, `ps_calc_10`, `ps_car_11` et `ps_calc_03`. Être un client risqué n'a pas d'impact sur la valeur de ces variables.






## Étude des dépendances (chi2, V de Cramer) entre la target et les variables explicatives catégorielles

```{r}
cramer = data.frame(NA, length(var_fac), 3)

for (i in (1:length(var_fac))){
    tab = table(train[,var_fac[i]], train$target)
    chisq_results = chisq.test(tab)
    cramer[i, 1] = var_fac[i]
    cramer[i, 2] = round(sqrt(chisq_results$statistic/(nrow(train))), 3)
    cramer[i, 3] = signif(chisq_results$p.value, 3)
}
colnames(cramer) = c("variable", "cramerv", "pvalue_chi2")
cramer %>% 
    arrange(-cramerv) %>% 
    ggplot() +
    aes(x = reorder(variable, -cramerv),
        y = cramerv,
        fill = -cramerv) +
    geom_bar(stat = "identity",
             show.legend = F) +
    xlab("Feature") +
    ylab("Cramer's V") +
    ggtitle("Cramer's V against the explained response") +
    theme(axis.text.x = element_text(angle = 45,
                                     hjust = 1)) +
    scale_fill_gradient(high = "pink",
                        low = "darkred")
```

## Étude des dépendances (chi2, V de Cramer) entre variables explicatives catégorielles

```{r}
cramer_cross = matrix(NA, length(var_fac), length(var_fac))

for (i in (1:length(var_fac))){
    for (j in (1:length(var_fac))){
        tab = table(train[, var_fac[i]], train[, var_fac[j]])
        chisq_results = chisq.test(tab)
        cramer_cross[i, j] = sqrt(chisq_results$statistic/(nrow(train) * (min(dim(tab)) -1)))
    }
}

cramer_cross = round(cramer_cross, 3)
colnames(cramer_cross) = var_fac
rownames(cramer_cross) = var_fac

corrplot(cramer_cross,
         method = "shade",
         type = "upper",
         diag = F,
         tl.srt = 45, 
         tl.col = "black",
         tl.cex = 0.6, 
         addCoef.col = "darkgreen", 
         addCoefasPercent = T)
```

Les variables catégorielles ne présentent pas de problème.

On ne modifie pas la liste des régresseurs.

# Exportations

```{r, eval = F}
levels(train$target) = c("no", "yes")
levels(test$target) = c("no", "yes")
type_col = sapply(train, class)
save(train, test, type_col,
     file = "./../data/bases_35features.RData")
```
